+++ workspace/node/bookmarks/src/bookmark/index.ts
@@ -1,3 +1,7 @@
 export * from "./node";
 export * from "./xbel";
 export * from "./xml";
+
+export const VIVALDI_SEPARATOR_HREF = "http://bookmark.placeholder.url/";
+export const VIVALDI_SEPARATOR_TEXT = "---";
+export const VIVALDI_SEPARATOR_DESCRIPTION = "separator";