+++ workspace/node/bookmarks/tests/parsers/netscape.test.ts
@@ -1,8 +1,12 @@
-import { readFile } from "node:fs/promises";
+import {
+    readFile,
+} from "node:fs/promises";
 
 import { describe, expect, it } from "vitest";
 
 import { NetscapeParser } from "@/parsers/netscape";
+
+import { deserializeObject } from "./utils";
 
 describe("netscape parser test", async () => {
     const parser = new NetscapeParser();
@@ -21,7 +25,18 @@ describe("netscape parser test", async () => {
             "XML objects",
         ).toEqual(bookmarks_example_html_xml_json);
 
-        /* TODO: XML parse result */
+        /* XML parse result */
+        const bookmarks = parser.parse(bookmarks_xml);
+
+        // console.log(JSON.stringify(bookmarks, null, 4));
+        // await writeFile("./data/bookmarks-example.html-node.json", JSON.stringify(bookmarks, null, 4));
+
+        const { default: bookmarks_example_html_node_json } = await import("~/data/bookmarks-example.html-node.json");
+        deserializeObject(bookmarks_example_html_node_json);
+        expect(
+            bookmarks,
+            "XML parse result",
+        ).toEqual(bookmarks_example_html_node_json);
     });
 
     it(`bookmarks-floccus.html`, async () => {
@@ -39,6 +54,17 @@ describe("netscape parser test", async () => {
             "XML objects",
         ).toEqual(bookmarks_floccus_html_xml_json);
 
-        /* TODO: XML parse result */
+        /* XML parse result */
+        const bookmarks = parser.parse(bookmarks_xml);
+
+        // console.log(JSON.stringify(bookmarks, null, 4));
+        // await writeFile("./data/bookmarks-floccus.html-node.json", JSON.stringify(bookmarks, null, 4));
+
+        const { default: bookmarks_floccus_html_node_json } = await import("~/data/bookmarks-floccus.html-node.json");
+        deserializeObject(bookmarks_floccus_html_node_json);
+        expect(
+            bookmarks,
+            "XML parse result",
+        ).toEqual(bookmarks_floccus_html_node_json);
     });
 });